Move widgets_container_impl.h to mainview folder.
